Pref. At. Simple: use bar pen for histogram
Unless there was a compelling reason for it, I think it is misleading to plot a histogram as a line plot. (When see power laws represented as line plots, those are usually CCDF plots, not histograms.)

There is also the matter of consistency with the original Preferential Attachement model, where the histogram is a bar plot.