Urban Suite - Tijuana Bordertowns: fix runtime error, close #61
After digging a bit deeper into this model, I realized that the whole functioning of this model is... unusual, to say the least. The setup is done from a forever button that ticks. Lots of stuff happens at specific tick values. There's a concept of "phases" that change as the setup progresses.

Not wanting to go down the rabbit hole, I went with the simplest change I could come up with to solve the problem: adding a global flag to indicate if the setup is completed or not.