Ant Lines: start leader with random heading, clarify setup (#103)
The leader's heading was previously reset by having `set heading 90` in the `ask turtles` block. I also removed `wiggle 50` which was doing nothing nothing (because a newly created turtle has a random heading anyway) and "centralized" `set size 2` in the `ask turtles` block.

I'd write the whole `setup` in a very different manner if I was starting anew but don't want to rock the boat too much on such an old model.