Tijuana Bordertowns: remove useless code
The model tried in many places to use `patch-right-and-ahead` and `patch-left-and-ahead` in addition to `patch-ahead`, but since the angle given is always 1, those calls always end up reporting the same patch as `patch-ahead`, making them redundant.

The original author probably meant to test patches on the left and right of the patch ahead and didn't realize that's not what was happening. Still, this commit preserves the model's actual behavior, even if it may not have been the originally intented one.