Tijuana Bordertowns: fix runtime exceptions
See https://github.com/NetLogo/models/issues/109#issuecomment-128519743 for problem description.

My approach is simply to kill the builders if the patch-ahead is nobody. It makes no sense for them to continue anyway since we're in a non-wrapping world.

I also added `patch-set` call in front of `one-patch patches in-radius 3 with [ ... ]` to avoid an "ask nobody" exception there.