Wolf Sheep Predation (docked): reduce dt to 0.001
Changing this in order to fix https://github.com/NetLogo/ccl-models/issues/19.

As explained in the issue, the old value of 0.1 generated increasing instability in the system dynamics part of the model. After discussion with @qiemem, we decided that it should be changed since:

- as far as we can see, there is no reason why it should be as high as 0.1;
- 0.001 is the value used in "Wolf Sheep Predation (docked)", and it makes sense for both models to use the same value.

(Note, however, that https://github.com/NetLogo/NetLogo/issues/650 is not resolved by this commit.)