Geometron Top-Down: remove wait (#140)
I removed the call to `wait 1`, which was (weirdly) blocking the whole model for full second anyway. To compensate, I reduced the potential amount of variation from 1.0 to 0.001 ("empirically" determined to give the same ballpark variation).

I went with this fairly minimal code change for #140, but I am under the impression that the whole "color shift" thing is not accomplishing what it aims to accomplish anyway. I may try to dig deeper into it when I tackle #145.