Traffic Intersection: modify light switching mecanism
The main goal was to get rid of `wait` in the "switch" button (#140), but the change goes deeper than that: it uses a tick countdown instead of trying to emulate one with `mod` (which wasn't really compatible with having a manual "switch" button.

I think that, overall, the new code is much easier to understand.