Traffic Intersection: prevent cars from running red lights
Fix #147.

It's not entirely clear to me what was happening, but I think it was an off by one error in the way `move` interpreted the result of the old `clear-ahead` procedure.

Rather than trying to fix this (imho) confusing code, I rewrote most of it in a way that is hopefully clearer and more robust.