Traffic Intersection: use turtle breeds for lights and accidents
Not only those this makes the code clearer, it also avoids interfering with the traffic lights when an accident occurs on the same patch as a light (since both concepts used to be represented by the pcolor of the patch).

Bonus: we get to use the "fire" shape for an accident, instead of a boring orange square.

I kept the lights as squares for consistency with other traffic models, even though I was tempted to use circles...

(#147)