allow `wait` in a bunch of ProbLab models
Some of the models rely on `wait` to achieve animation effects that would be hard to achieve otherwise. Another pattern is to wait at certain tick intervals: a few ticks happen, then the model waits to let the user see the result.

Perhaps, with more work, some usages of `wait` could be eliminated, but I'm generally reluctant to touch the ProbLab models: they're generally complicated and somewhat brittle. "Random Combinations and Permutations" is a good example of that...