use parallel collection to test for buttons causing runtime errors
I had to turn off the bytecode generator in order for this to work: for some
reason, it was triggering  a weird exception:

java.lang.ClassCastException:
org.nlogo.prim._asm_proceduresetup_constdouble_0
cannot be cast to org.nlogo.nvm.Instruction

No idea what that means. But I've seen that before when experimenting
with testing models in parallel. Now that I know that turning off the
generator avoids it, I could revisit other tests to see if there are speed
gains to be made...