Wealth Distribution: add "If not, why not?" to a question in the info tab...
The info tab asks "Try allowing metabolism or vision or another
characteristic to be inherited.  Will we see any sort of evolution?"

In the model, each agent has one and only one offspring, regardless of how
well they're doing, so one of the main conditions for evolution (differential
reproduction) is not met. It doesn't mean that the user shouldn't try it
anyway, but it seems safer (kinder?) to me to at least nudge them into
thinking that it *may* not work...