Paths: simplify patch popularity management
By allowing popularity to go below 1, we simplify the code: since popularity
decays by some percentage each, it will never go below 0, so there is no
need to check for that. And it makes no behavioral difference: I've changed
the check for minimum popularity from > to >=, to compensate (which
makes more sense to me anyway).

The other difference is that, since buildings are now turtles, there is no need
to treat the patches underneath buildings differently: we let their popularity
increase and decay just like that of any other patch.