shorten lines in various models (close #111)
Ideally, I'd like to bring all models to an 85 characters limit (because that's
the limit for IABM textbook models, given MIT Press' page layout) but
that would take considerable effort.

So, for now, I adopted a limit of of 170 characters. That's still on the
longuish side in my opinion, but that's noentheless a significant
improvement from what some models had before.