LS Examples: Update Preview Images
I didn't realize you had to actually go into each model and specify the
command (I think that's odd / maybe stupid). But it's fixed now.

So in order to add a manual preview image, you have to: 1. Create a
preview image; 2. Add that preview image within NetLogo; 3. Add that
model to the Exceptions list. That seems like a lot...