Info Tab Example: keep notarizer test happy
This is a subtle one. We have a test that checks that every model file is
identical to the output of the notarizer running on that file. When the
notarizer writes back a notarized file, it adds a blank line after first and
second level headings. The Info Tab Example has "fake" headings shown
in a "text" code block.  These weren't a problem before because they were
indented by four spaces and the notarizer didn't see them as headings. But
now that they are included in a ``` and aligned to the left margin, it does.

The added blank lines keeps everyone happy.