Fireworks: Change `fireworks` to `max-fireworks`
In order to address #340, it seems the best change is to rename the
`fireworks` parameter. The original model selects a random number of
fireworks between [0, `fireworks`). This means that with `fireworks =
1`, there are no fireworks and at any value of `fireworks` you can have
a round with 0 fireworks.

This update addresses those two things by:

1. Rename `fireworks` to `max-fireworks`. Because of the randomness
selected by the model author, this more accurately conveys to the user
why they may not see exactly the number of `fireworks` go off that they
selected.

2. Enforce a strict minimum of 1 firework (by using `(random
max-fireworks) + 1`. This also means that when you select a
`max-fireworks` of 18, you can actually see 18 fireworks. In the
previous version, you would see at most 17.