AIDS: Rename to HIV
As per a request by a NetLogo-er, we've decided to rename the `AIDS`
model to `HIV`. This is because HIV is a transmittable virus as the
model explains while AIDS is not and cannot, in fact, be transmitted.
This commit addresses the renaming of the model in terms of the Models
Library file structure causing the following changes:

+ A renaming of `AIDS.nlogo` to `HIV.nlogo`
+ An update of the CrossReferencer
+ An update of the gitignore
+ An update to exemptions for CodeTests