Tumor: Small updates to fix #386
Fixes #386 reported by Ken Kahn. This error was the result of a request (via a button) to kill a turtle with a specific `who` number. As such, if that turtle does not exist when the button is pressed (for example, on the second press if the button is pressed twice, the model will throw this error. _This is expected behavior._ However, I can see a case where the user has either A) forgotten that they already pressed the button or B) clicks the button twice in quick succession by accident. As such, I've added two "try-catch" blocks to each of the buttons that use this functionality of asking specific turtles to `die`. This avoids the error throwing that was reported.

Additionally, small style updates are included to match the newer guidelines.