ANN Perceptron: Update thickness scaler
Bumping up the thickness scaling factor in the perceptron case provides better uniformity between this model and the mulit-layer model. In that model, the weights end up in a larger range of values, so you can see the thickness of the connections change. Because of the rather limited range of weights in this model, we need a higher scaling factor in order to actually see the change in thickness over time. This should resolve NetLogo/NetLogo#1704.