Bug Hunt Preadtors and Invasive Species: Fix eat ordering
This commit fixes #391.

In previous versions of the model, `bugs` were able to eat _before_ `invaders` which means that `bugs` had a _slight_ advantage in getting the most tasty grass over `invaders`. We fix that by reworking the main `go` procedure so that `bugs` and `invaders` are given the chance to eat in random order. This *does* affect the outcome of the model, as it is rectifying a  "bias" of sorts towards the `bugs`. New behavior presents a better balance between `bugs` and `invaders` "wins" as the final outcome of the model.

This updates two slightly different versions of the `Bug Hunt Predators and Invasive Species` Model.

1. BEAGLE Curriculum - Bug Hunt Predators & Invasive Species
2. ModelSim - Bug Hunt Predators & Invasive Species - Two Regions