Peppered Moths: Increase moth size
On higher resolution screens, the default size of the `moth` shape looks like a square. To fix this, we make the moths slightly bigger. Because no part of the model's mechanisms rely on the sizes or collisions (e.g. mating is asexual; death is random), this is simply a display change.