Update to fix tests on Jenkins
This temporarily disables automatic testing of the `Profiler Example` model due to an issue with the LS extension turning off the bytecode generator. Due to the order in which the models are processed, this means that the LS model tests essentially disable the Profiler model tests. For now, we'll locally test the Profiler Example model until a non-invasive solution is found.