Make `onServer` boolean to actually work
This allows us to disable certain tests on Jenkins vs locally. I'm sure there's some way to make the previous detection certainly work, but this is such an exceptional case (there's only two tests that come to mind that shouldn't be run always) that we've changed to an isomorphic way of specifying local testing as we do for 3d testing.

If you want to run local-only tests (as of this commit only InfoTabURL tests and some hard-to-compile-on-the-server models), you just need to specify the `org.nlogo.onLocal` system property as `true`. This is similar to how we use `org.nlogo.is3d` to run tests on 3d models using 3d headless.

So in summary, to run local tests: `sbt -Dorg.nlogo.onLocal=true`. All other testing commands remain exactly the same.