Add prototype NLW Compatability Test
This is really just a placeholder for more advanced tests that determine the viability of a library model for NLW. Right now, we just check for prims that have yet to be implemented in NLW. On run on local since it's not a helpful failure. Update these occasionally from some authoritative source.

I've been using https://github.com/NetLogo/Tortoise/wiki/Unimplemented-Primitives