Update to recognize new CrossReference style
So in PR #362, we changed the way models are Cross Referenced in the Library. Before, they were literallly copied into the appropriate folders on release. But now, they just "appear" in the library index in multiple places rather than having two copies of each file.

We apparently never updated the .gitignore to reflect that leaving a bunch of stuff that's totally not necessary chilling in here. This is the first commit to try to rectify that.