Notarizer: Fix HubNet detection.
This fixes HubNet "detction" for models notarizing. The old method only marked as HubNet models, ones that were in a folder called "HubNet Activities." This works fine...most of the time. But if a HubNet model was, say, in a Curricular Models folder and NOT in its own sub-folder, it would just miss the HubNet citation all together. That shouldn't be the case.

Interestingly, it also had the effect of sometimes double tagging HubNet models via their `compressed name`. For instance, `HubNetPrisonersDilemmaHubNet`. Luckily the only thing that uses that (the NetLogo/netlogo-site cgi files) seems to ignore the pre-pend so any changes here don't neccesitate changes in that repo.

I'm fairly confident that this now solves ALL our HubNet citation issues...iff that HubNet models are named acccording to the Style Guide: `<model-name> HubNet`.