Add Materials Science folder inside Chemistry & Physics
Our Chemistry & Physics folder now contains enough examples that might be classified as "Materials Science" that they deserve their own folder. This means a number of models have been physically moved into this new sub-folder of Chemistry and Physics:
- Solid Diffusion
- Polymer Dynamics
- Lennard-Jones
- Dislocation Motion and Deformation
- Crystallization Models (3)

Also in this folder are the MaterialSim models (only 1). To do this, we've moved MaterialSim into the Curricular Models folder and then crossreferenced its contents to the Materials Science folder.